cli: Fix build with slibtool.
The .libs directory is an internal libtool directory which build
projects should not directly use. With libtool it will happen to work
because it parses the actual file, but with other implementations such
as slibtool it will fail because it relies a .deps file found in the
same directory as the .la file which will not be found.

Instead of using the the symlinked .la file in the .libs directory its
better to use the file the symlink points to and this will allow greater
portability for building WavPack.