Add CI
Adds CI which runs `make distcheck`, automatically creates the tarball and Windows binaries. It is triggered on push, pull requests and release, on release it will attach the tarball and the zipped windows binaries to the release.
The Windows zip files includes all files that were installed, but this can be adjusted if desired.