Fix autoreconf with gettext 0.25
Since gettext 0.25, the AM_ICONV macro is installed into the m4
directory, so we need to add that autoconf config directory.  We also
need to make sure autoreconf recognizes that it needs to run autopoint
to install the gettext macros, which it does heuristically by checking
whether AM_GNU_GETTEXT_VERSION is used, so add that and the minimum
amount of other stuff required to make it happy.  The version of
gettext we say we use shouldn't matter, so I just went with the
version in ubuntu:latest since that's what CI uses.